backend/cloud_functions/data_refresh/requirements.txt
google-cloud-storage
google-cloud-secret-manager
# dotenv